extract commit and proposal from crypto-block package
