+++ packages/contracts/source/contracts/crypto/index.ts
@@ -1,5 +1,6 @@
 export * from "./block";
 export * from "./config";
+export * from "./commit";
 export * from "./crypto";
 export * from "./enums";
 export * from "./hash";
