+++ packages/contracts/source/contracts/crypto/networks.ts
@@ -1,4 +1,4 @@
-import { CommittedBlockJson } from "./block";
+import { CommittedBlockJson } from "./commit";
 
 export type NetworkConfig = {
 	genesisBlock: CommittedBlockJson;
