+++ packages/core/package.json
@@ -42,6 +42,7 @@
 		"@mainsail/crypto-address-keccak256": "workspace:*",
 		"@mainsail/crypto-address-ss58": "workspace:*",
 		"@mainsail/crypto-block": "workspace:*",
+		"@mainsail/crypto-commit": "workspace:*",
 		"@mainsail/crypto-config": "workspace:*",
 		"@mainsail/crypto-consensus-bls12-381": "workspace:*",
 		"@mainsail/crypto-hash-bcrypto": "workspace:*",
