+++ packages/crypto-block/source/serializer.test.ts
@@ -75,21 +75,4 @@ describe<{
 		assertCommitData(assert, deserialized, commit);
 	});
 
-	it("#serialize - should serialize and deserialize lock proof", async ({ deserializer, serializer }) => {
-		const proposalLockProof: Contracts.Crypto.AggregatedSignature = {
-			signature:
-				"927628d67c385fe216aa800def9cce0c09f5f9fbf836583d7c07ab6a98e1b5681802c92f81ad54984236a07fa389dbab1519f3c91ad39a505a61c3624a88c65da71fe721d7af0ed452516771b94d027be713dba68e14fa2c9680e35b63f0e038",
-			validators: [true, true, true, false, false, true, true, true, true, false],
-		};
-
-		const serializedProposalLockProof =
-			"927628d67c385fe216aa800def9cce0c09f5f9fbf836583d7c07ab6a98e1b5681802c92f81ad54984236a07fa389dbab1519f3c91ad39a505a61c3624a88c65da71fe721d7af0ed452516771b94d027be713dba68e14fa2c9680e35b63f0e0380ae701000000000000";
-
-		const serialized = (await serializer.serializeLockProof(proposalLockProof)).toString("hex");
-		assert.equal(serialized, serializedProposalLockProof);
-
-		const deserialized = await deserializer.deserializeLockProof(Buffer.from(serialized, "hex"));
-		assert.equal(proposalLockProof.signature, deserialized.signature);
-		assert.equal(proposalLockProof.validators, deserialized.validators);
-	});
 });
