+++ packages/crypto-commit/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "distribution"
+	},
+	"include": ["source/**/**.ts"]
+}
