+++ packages/p2p/source/peer-verifier.ts
@@ -25,8 +25,8 @@ export class PeerVerifier implements Contracts.P2P.PeerVerifier {
 	@inject(Identifiers.Database.Service)
 	private readonly database!: Contracts.Database.DatabaseService;
 
-	@inject(Identifiers.Cryptography.Block.Factory)
-	private readonly blockFactory!: Contracts.Crypto.BlockFactory;
+	@inject(Identifiers.Cryptography.Commit.Factory)
+	private readonly blockFactory!: Contracts.Crypto.CommitBlockFactory;
 
 	@inject(Identifiers.P2PLogger)
 	private readonly logger!: Contracts.P2P.Logger;
@@ -85,7 +85,7 @@ export class PeerVerifier implements Contracts.P2P.PeerVerifier {
 		}
 
 		// TODO: Support header only requests
-		const receivedCommittedBlock = await this.blockFactory.fromCommittedBytes(blocks[0]);
+		const receivedCommittedBlock = await this.blockFactory.fromBytes(blocks[0]);
 
 		const blockToCompare =
 			block.data.height === heightToRequest ? block : await this.database.getBlock(heightToRequest);
